requirements.txt
kani-tts
transformers==4.57.1
datasets
soundfile
pyyaml
huggingface_hub
numpy
